Split AndroidSampler into UserAgentSampler, WifiSampler and TelephonySampler
